Update VDI_By_Day_Compute_By_Night_v0.1.ps1
Added the ability to start at a given VM, so lets say VMs 001 & 002 need to always be on and dedicated to workload application (a controller for example) you can now set that starting VM so it will skip VMs 001 & 002. Also added some formatted messaging so it was clearer what was happening.